requirements.txt
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037097816/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
async-lru==1.0.2
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///home/ktietz/src/ci_mi/defusedxml_1612808095953/work
entrypoints==0.3
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1612212923938/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213311222/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
MarkupSafe==1.1.1
mistune==0.8.4
nbclient @ file:///tmp/build/80754af9/nbclient_1613676291033/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914830498/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///tmp/build/80754af9/notebook_1611340975709/work
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120460739/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyaes==1.6.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
Pyrogram==1.1.13
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141720057/work
PySocks==1.7.1
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1612458529756/work
QtPy==1.9.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1605205327372/work
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
TgCrypto==1.2.2
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
Wand==0.6.6
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
